Fix effective counting for bubbleGraph, split graph utils
